refactor: move ignore code into _ignores.py
